Move Throttler to separate file
